Latest rolling window code for enacting cooling control
A large file - have a look after the section labeled `BREAKPOINT` for how went about calling the cooling.exe file as well as preparing the intermediate files.  Above the BREAKPOINT there is a time of day based controller and a code segment where I use Bokeh plot for generating a seperate html file for data visualization and analysis. Please excuse the mess.